Test - Camera Movement Sorted for Left & Right
So, after a bit of tinkering I was able to find a solution that allowed the camera to move around the target object without stuttering or snapping back to the camera movement position. I will now attempt to get up and down working as well and maybe have a ramp up of some kind so the camera can move fully around the target in 360(degrees) freely without running out of room to move the camera.