Test - Modelling & Texturing Ground Assets #1
I've made a basic ground asset with and without walls to see how Unity handles multi sub textured objects. In short not well. Its easier to make a texture map and texture the asset in 3DS before export it to ensure the asset looks as intended. Might also make the assets either 100x100 or 200x200 in 3DS for scale sake, we'll see. The current assets are 150x150 with 2048 texture maps