Prep folder structure & Gems
I've added Gems to the game, the V2 is better that the V1 for size. Preped the folder structre for later assets and fixed there been 2 locations for textures. Also tweaked the materials for the start and finish.

Another universal script is in this too, rotate script which rotates as a chosen speed on a chosen axis.