Back at it, FIxed Jump a bit 'n' stuff
Having a little more of a go at this for this week, just to get a menu & level start and end state made at-least.

I've somewhat fixed the jump, its still a tad weird when going backwards but otherwise it works. I've also added a new script that reloads the level you are on (only for debug purposes).